requirements.txt
folium==0.11.0
geopy==1.22.0
python-dotenv==0.13.0
requests~=2.2.0
Flask~=1.1.2